gc: a program can open 63 regions
`JL_GC_MAX_REGIONS` goes from 8 to 64: region 0 stays the stock heap, and a
program can open regions 1 to 63. Two things bounded the count. The constant
itself, and `region_quarantined_mask`, a 32-bit word: its `fetch_or` in
`jl_gc_region_wb` shifted a 32-bit one by the region number, so it widens to
64 bits. Everything else held 64 already. The page tag, the task field, the
parent table and the child counts are bytes; `region_live_mask`,
`region_haschild_mask` and `region_uptree[]` are 64-bit words, and bit 63
is region 63.

No hot path reads the count. The allocation fast path reads `active_pools`,
the barrier reads two page tags and one uptree word. The three brackets of a
stock collection skip 64 entries per heap instead of 8, and a thread heap
carries 64 region entries in place of 8.

The scripts share `MAX_REGIONS` in `regions_api.jl`, and the three that used
8 as a bad region number use it. A new script, `regions_many.jl`, opens a
window on every region, makes the legal stores at the top of the default
chain, resets the chain from 63 down to 1 after the parents refused with
ECHILD, declares a tree of 62 leaves under one trunk and resets it, and
quarantines region 32 and region 63: the bit a 32-bit mask lacked, and the
top bit. The ten scripts pass at every configuration of the harness.

+++ src/gc-regions.c
@@ -42,7 +42,7 @@ static _Atomic(int) region_windows_open = 0;
 // child's region is quarantined - its reset and census refuse from then on,
 // so an escape costs memory, never a dangling pointer.
 JL_DLLEXPORT _Atomic(uint8_t) jl_gc_region_barrier_on = 0;
-static _Atomic(uint32_t) region_quarantined_mask = 0;
+static _Atomic(uint64_t) region_quarantined_mask = 0;
 
 // The census filter: the region whose census runs now, 0 otherwise. The mark
 // loop reads it once per object array and passes it down, so a stock mark
@@ -132,8 +132,8 @@ JL_DLLEXPORT int jl_gc_region_declare_parent(int child, int parent)
     if (jl_atomic_load_relaxed(&region_windows_open) != 0)
         return JL_GC_REGION_EBUSY;
 
-    // The world stops for the test and the rebuild together. The eight
-    // uptree words change one at a time, so a thread that opened a window
+    // The world stops for the test and the rebuild together. The uptree
+    // words change one at a time, so a thread that opened a window
     // between a test and a rebuild done apart could read a half-built tree
     // and judge a store against it. A declaration is a startup act and runs
     // a handful of times, so the pause costs nothing that matters.
@@ -247,8 +247,8 @@ JL_DLLEXPORT void jl_gc_region_wb(const void *parent, const void *child) JL_NOTS
     int cr = 0, pr = 0;
     if (__likely(!region_store_escapes(parent, child, &cr, &pr)))
         return;
-    uint32_t bit = (uint32_t)1 << cr;
-    uint32_t seen = jl_atomic_fetch_or_relaxed(&region_quarantined_mask, bit);
+    uint64_t bit = (uint64_t)1 << cr;
+    uint64_t seen = jl_atomic_fetch_or_relaxed(&region_quarantined_mask, bit);
     if (!(seen & bit))
         jl_safe_printf("REGION-ESCAPE: a %s of region %d was stored into a %s "
                        "of region %d; region %d is quarantined - its reset and "
